limite marche pas même erreur

+++ SmashBigBordelv2/Jeu.h
@@ -72,9 +72,9 @@ protected:
 
 	std::vector < sf::Drawable* > vectorHUD;
 
-	const int VitesseDeplacement = 150;//  pixel/sec
-	const int VitesseSaut = 20;
-	const int Gravity = 9.81;
+	const float VitesseDeplacement = 150;//  pixel/sec
+	const float VitesseSaut = 20;
+	const float Gravity = 9.81;
 
 	const float Spawn_x1 = 100;
 	const float Spawn_y1 = -500;
